fix: implement select-before-modify constraint for metadata and fix MFT engine bugs
- Implemented "must select before modify" constraint in Grid, Thumbnail, and Tree delegates (Plan #38).
- Fixed USN V3 record offset error in MftReader for Win11/ReFS compatibility.
- Resolved iterator invalidation and deadlock risks in MftReader by optimizing lock usage and signal emission.
- Refactored MftReader signals (entryAdded/entryUpdated) to use 32-bit indices for better performance.
- Fixed SyncEngine's full scan logic and real-time monitoring restoration after cache load.
- Added fallback counting in CategoryRepo to handle empty database scenarios (Plan #39).

+++ src/ui/ScanDialog.cpp
@@ -763,7 +763,7 @@ ScanDialog::ScanDialog(QWidget* parent)
     });
 
     // 2026-05-28 核心补丁：监听引擎增量信号，实现标题栏计数实时更新
-    connect(&MftReader::instance(), &MftReader::entryAdded, this, [this](uint64_t) { updateStatus("就绪"); });
+    connect(&MftReader::instance(), &MftReader::entryAdded, this, [this](uint32_t) { updateStatus("就绪"); });
     connect(&MftReader::instance(), &MftReader::entryRemoved, this, [this](uint64_t) { updateStatus("就绪"); });
 
     // 2026-05-16 物理重载：断开基类 Qt 置顶逻辑，改用 Win32 原生 SetWindowPos 以实现无损切换
